Moved Page into own directory with all components split for better readablility
